:recycle: Relaxation limit of candidate type
