+++ search.ts
@@ -3,16 +3,11 @@ import { bitDP } from "./bitDP.ts";
 
 export interface Candidate {
   title: string;
-  titleLc: string;
-  updated: number;
-  metadata: {
-    project: string;
-    hasIcon: boolean;
-  }[];
 }
-export interface CandidateWithPoint extends Candidate {
+export interface MatchInfo {
   /** 編集距離 */
   dist: number;
+
   /** queryがマッチした位置
    *
    * - 1番目：開始位置
@@ -34,13 +29,13 @@ const getMaxDistance = [
 ];
 
 /** 一致する候補をしぼりこむ函数*/
-export interface Filter {
+export interface Filter<T extends Candidate> {
   /** 一致する候補をしぼりこむ函数
    *
    * @param source 検索候補リスト
    * @return 一致した候補
    */
-  (source: readonly Candidate[]): CandidateWithPoint[];
+  (source: readonly T[]): (T & MatchInfo)[];
 }
 
 /** `query`に曖昧一致する候補を絞り込む函数を作る
@@ -48,7 +43,9 @@ export interface Filter {
  * @param query 検索語句
  * @return 検索函数。検索不要なときは`undefined`を返す
  */
-export const makeFilter = (query: string): Filter | undefined => {
+export const makeFilter = <T extends Candidate>(
+  query: string,
+): Filter<T> | undefined => {
   /** キーワードリスト
    *
    * - 空白は取り除く
@@ -66,15 +63,15 @@ export const makeFilter = (query: string): Filter | undefined => {
     for (const query of queries) {
       result = filter(query, max, result);
     }
-    return result as CandidateWithPoint[];
+    return result as (T & MatchInfo)[];
   };
 };
 
-const filter = (
+const filter = <T extends Candidate>(
   query: string,
   maxDistance: number,
-  source: (Candidate & { dist?: number; matches?: [number, number][] })[],
-): CandidateWithPoint[] => {
+  source: (T & Partial<MatchInfo>)[],
+): (T & MatchInfo)[] => {
   const m = [...query].length;
   const filter_ = bitDP(query);
 
@@ -105,51 +102,7 @@ const filter = (
       if (newDist > maxDistance) return [];
 
       matches.push([newMatch.start, newMatch.end]);
-      return [{ title, dist: newDist, matches, ...props }];
+      return [{ title, dist: newDist, matches, ...props } as (T & MatchInfo)];
     },
   );
-};
-
-/** 候補を並び替える
- *
- * @param candidates 並び替えたい候補のリスト
- * @param projects projectの優先順位付けに使う配列。優先度の高い順にprojectを並べる
- * @return 並び替え結果
- */
-export const sort = (
-  candidates: readonly CandidateWithPoint[],
-  projects: readonly string[],
-): CandidateWithPoint[] => {
-  const projectMap = Object.fromEntries(
-    projects.map((project, i) => [project, i]),
-  );
-
-  return [...candidates].sort((a, b) => {
-    // 1. 編集距離が短い順
-    const diff = a.dist - b.dist;
-    if (diff !== 0) return diff;
-
-    // 2. マッチ位置が早い順
-    const sa = a.matches.map(([s]) => s).sort();
-    const sb = b.matches.map(([s]) => s).sort();
-    for (let i = 0; i < sa.length; i++) {
-      const sdiff = sa[i] - (sb[i] ?? sb.length);
-      if (sdiff !== 0) return sdiff;
-    }
-
-    // 3. 文字列が短い順
-    const ldiff = a.title.length - b.title.length;
-    if (ldiff !== 0) return ldiff;
-
-    // 4. projectsで若い順
-    const pdiff = Math.min(
-      ...a.metadata.map((meta) => projectMap[meta.project] ?? projects.length),
-    ) - Math.min(
-      ...b.metadata.map((meta) => projectMap[meta.project] ?? projects.length),
-    );
-    if (pdiff !== 0) return pdiff;
-
-    // 5. 更新日時が新しい順
-    return b.updated - a.updated;
-  });
 };
