+++ source.ts
@@ -1,0 +1,9 @@
+export interface Candidate {
+  title: string;
+  titleLc: string;
+  updated: number;
+  metadata: {
+    project: string;
+    hasIcon: boolean;
+  }[];
+}
